% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute1ValperPeak.R
\name{compute1ValperPeak}
\alias{compute1ValperPeak}
\title{Title}
\usage{
compute1ValperPeak(bed, wig, w = 20000, seqlens, fun = "sum")
}
\arguments{
\item{bed}{A GRanges object}

\item{wig}{A RLEList object}

\item{w}{Integer used to extend bed of both side : 2*w}

\item{seqlens}{A list wich contain the chromsize and their names}

\item{fun}{A character, both mean and sum are accepted}
}
\value{
A vector of integer for the coverage of each element in bed
}
\description{
Title
}
