% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeProfile.R
\name{computeProfile}
\alias{computeProfile}
\title{computeProfile : Get the average/sum coverage of a RleList object for a list of genomic positions}
\usage{
computeProfile(bed, wig, w = 20000, span = 200, seqlens,
  method = "mean")
}
\arguments{
\item{bed}{A GRanges object}

\item{wig}{A RLEList object}

\item{w}{Integer used to extend bed of both side : 2*w}

\item{span}{Integr used to determine the size of each bin of our 2*w interval}

\item{seqlens}{A list wich contain the chromsize and their names}

\item{method}{A character, both mean and sum are accepted}
}
\value{
A matrix of size length(bed)*(w/span)
}
\description{
computeProfile : Get the average/sum coverage of a RleList object for a list of genomic positions
}
