% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeProfileGenes.R
\name{computeProfileGenes}
\alias{computeProfileGenes}
\title{Title}
\usage{
computeProfileGenes(bed, wig, w = 3000, span = 200, seqlens,
  fun = "sum", get = "one")
}
\arguments{
\item{bed}{A GRanges object}

\item{wig}{A RLEList object}

\item{w}{Integer used to extend bed of both side : 2*w}

\item{span}{Integr used to determine the size of each bin of our 2*w interval}

\item{seqlens}{A list wich contain the chromsize and their names}

\item{fun}{A character, both mean and sum are accepted}

\item{get}{either "one" or "all". If get == "one", perform fun, if not, return all values}
}
\value{
If get == "one", return a vector of value for each position w/span, if get =="all", return a matrix with each values.
}
\description{
Title
}
