% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Get1val.R
\name{Get1val}
\alias{Get1val}
\title{Get1val : Get the coverage of a RleList object for a list of genomic positions}
\usage{
Get1val(Name, one.w, x)
}
\arguments{
\item{Name}{The name of the bigwig file}

\item{one.w}{A bigwig file imported in RleList format (as = "RleList")}

\item{x}{A GRanges object}
}
\value{
A tibble with the coverage of each element of x
}
\description{
Get1val : Get the coverage of a RleList object for a list of genomic positions
}
