% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetBWList.R
\name{GetBWList}
\alias{GetBWList}
\title{Get List of all bigwig}
\usage{
GetBWList(mydir = "/mnt/NAS/")
}
\arguments{
\item{mydir}{"/mnt/NAS/"}
}
\value{
A list of existing BW
}
\description{
Get List of all bigwig
}
