% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetDE_IR_GSE110386.R
\name{GetDE_IR_GSE110386}
\alias{GetDE_IR_GSE110386}
\title{Get RNA-Seq table DE results for IR GSE110386.}
\usage{

  GetDE_IR_GSE110386(path = "/home/rochevin/Documents/PROJET_THESE/DSB_EFFECT_ON_GENE/data/DDR_GSE110386/counts",
  pattern = "_counts.tsv", exclude = NULL,
  Conditions = paste(rep("WT", 9), rep(c("ctr", "4Gy4h", "10Gy4h"), each
  = 3), sep = "_"), Labels = paste(Conditions, rep(c("1", "2", "3"), 3),
  sep = "_"), p.cutoff = 0.1, LogFC.cutoff = 1,
  to.test = "WT_Cvs4Gy")
}
\arguments{
\item{path}{The path directory of table counts matrix (from HTSeq Count)}

\item{pattern}{The pattern of the extension (default "_counts.tsv")}

\item{exclude}{An regex expression to exclude some conditions, default NULL}

\item{Conditions}{A vector of conditions names}

\item{Labels}{The names of columns, default}

\item{p.cutoff}{The adjusted p.value cutoff used to determine wich gene is differentially expressed}

\item{LogFC.cutoff}{The log2(foldchange) cutoff used to determine wich gene is differentially expressed}

\item{to.test}{Which condition to test ? Default "CvsC_OHT"}
}
\value{
A data.frame with all expressed genes in this Condition
}
\description{
Get RNA-Seq table DE results for IR GSE110386.
}
