% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetMotifFROMJASPAR.R
\name{GetMotifFROMJASPAR}
\alias{GetMotifFROMJASPAR}
\title{Title}
\usage{

  GetMotifFROMJASPAR(bed_path = "/home/rochevin/Documents/PROJET_THESE/DSB_EFFECT_ON_GENE/results/COMMON_GDR_RESPONSE/BED",
  pattern = ".bed", species = 9606, collection = "CORE",
  myJASPAR = JASPAR2018::JASPAR2018,
  genome = BSgenome.Hsapiens.UCSC.hg19::BSgenome.Hsapiens.UCSC.hg19,
  w = 3000)
}
\arguments{
\item{bed_path}{A path which contains one or more bedfiles}

\item{pattern}{extension, default ".bed"}

\item{species}{The species ID from NCBI, default 9606}

\item{collection}{collection=c("CORE", "CNE", "PHYLOFACTS", "SPLICE", "POLII", "FAM", "PBM", "PBM_HOMEO", "PBM_HLH")}

\item{myJASPAR}{A JASPAR connection databases, default JASPAR2018::JASPAR2018}

\item{genome}{A BSgenome object}

\item{w}{upstream/downstream for promoter, default 3000}
}
\value{
A data.frame of motif
}
\description{
Title
}
