% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeFisherForVenn.R
\name{MakeFisherForVenn}
\alias{MakeFisherForVenn}
\title{Fisher Test for a list of name}
\usage{
MakeFisherForVenn(venn, universe)
}
\arguments{
\item{venn}{A list of names}

\item{universe}{Universe of names}
}
\value{
A vector of pval by fisher.test
}
\description{
Fisher Test for a list of name
}
