% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoveDamagedGenes.R
\name{RemoveDamagedGenes}
\alias{RemoveDamagedGenes}
\title{Remove Damaged Genes from table}
\usage{
RemoveDamagedGenes(my.vals,
  DSBpath = "/mnt/NAS/DATA/AsiSI/ASIsites_hg19_174_clived_IQ1.5.bed",
  GenePath = "/home/rochevin/Documents/PROJET_THESE/DSB_EFFECT_ON_GENE/data/EnsDB.Hsapiens.v75.genes.bed")
}
\arguments{
\item{my.vals}{A data.frame or tibble wich contains Gene id as rowname}

\item{DSBpath}{Path of a bedfile wich contains genomic position to exclude}

\item{GenePath}{Path of Gene annotation (ensembl format)}
}
\value{
my.vals without Genes that overlap with DSBpath
}
\description{
Remove Damaged Genes from table
}
